hw06/requirements.txt
Django==2.2.3
Pillow==6.1.0
coreapi==2.3.3
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.8.3
uwsgi==2.0.18